feat: rm types into types folder
