+++ packages/core/src/providers/provider-jsonrpc-tonx.ts
@@ -1,516 +1,45 @@
-import { Address } from "@ton/core";
 import { CreateAxiosDefaults } from "axios";
 import { z } from "zod";
 import { RunAction } from "./abstract-provider";
 import { Network } from "~core/types/network";
 import HttpFetchClient from "~core/utils/http-fetch-client";
 import { JsonRpcProvider } from "./provider-jsonrpc";
-import type { JsonRpcApiProviderOptions } from "./provider-jsonrpc";
+import {
+  GetTransactionsParams,
+  GetJettonBurnsParams,
+  GetJettonMastersParams,
+  GetJettonTransfersParams,
+  GetJettonWalletsParams,
+  GetMessagesParams,
+  GetNftCollectionsParams,
+  GetNftItemsParams,
+  GetNftTransfersParams,
+  EstimateFeeParams,
+  GetBlockHeaderParams,
+  GetBlockTransactionsParams,
+  RunGetMethodParams,
+  RunGetMethodV3Params,
+  RunGetMethodResponse,
+  GetTgBTCBalanceParams,
+  GetTgBTCHoldersParams,
+  GetTgBTCBurnsParams,
+  GetTgBTCWalletAddressByOwnerParams,
+  RadixConversionParams,
+  BinaryConversionParams,
+  GetTgBTCTransferPayload,
+  GetTgBTCBalanceResponse,
+  GetTgBTCBurnsResponse,
+  GetTgBTCConfigResponse,
+  GetTgBTCHoldersResponse,
+  GetTgBTCMasterAddressResponse,
+  GetTgBTCTransferResponse,
+  GetTgBTCTransfersParams,
+  GetTgBTCTransfersResponse,
+  GetTgBTCWalletAddressByOwnerResponse,
+  TgBTCMetaDataResponse,
+} from '~core/types/types';
+import { TONXJsonRpcProviderOptions, TONXRunAction } from "~core/types/action-types";
 
-type GetAccountBalanceParams = {
-  address: string;
-};
-
-type GetTransactionsParams = {
-  account?: string;
-  end_lt?: number;
-  end_utime?: number;
-  hash?: string;
-  limit?: number;
-  offset?: number;
-  seqno?: number;
-  shard?: string;
-  sort?: string;
-  start_lt?: number;
-  start_utime?: number;
-  workchain?: number;
-};
-
-type GetJettonBurnsParams = {
-  address?: string;
-  end_lt?: number;
-  end_utime?: number;
-  jetton_master?: string;
-  jetton_wallet?: string;
-  limit?: number;
-  offset?: number;
-  sort?: "ASC" | "DESC";
-  start_lt?: number;
-  start_utime?: number;
-};
-
-type GetJettonMastersParams = {
-  address?: string;
-  admin_address?: string;
-  limit?: number;
-  offset?: number;
-};
-
-type GetJettonTransfersParams = {
-  address?: string;
-  direction?: "in" | "out" | "both";
-  end_lt?: number;
-  end_utime?: number;
-  jetton_master?: string;
-  jetton_wallet?: string;
-  limit?: number;
-  offset?: number;
-  sort?: "ASC" | "DESC";
-  start_lt?: number;
-  start_utime?: number;
-};
-
-type GetJettonWalletsParams = {
-  address?: string;
-  jetton_address?: string;
-  limit?: number;
-  offset?: number;
-  owner_address?: string;
-};
-
-type GetMessagesParams = {
-  body_hash?: string;
-  destination?: string;
-  hash?: string;
-  limit?: number;
-  offset?: number;
-  source?: string;
-};
-
-type GetNftCollectionsParams = {
-  collection_address?: string;
-  limit?: number;
-  offset?: number;
-  owner_address?: string;
-};
-
-type GetNftItemsParams = {
-  collection_address?: string;
-  limit?: number;
-  offset?: number;
-  owner_address?: string;
-};
-
-type GetNftTransfersParams = {
-  address?: string;
-  collection_address?: string;
-  direction?: "in" | "out" | "both";
-  end_lt?: number;
-  end_utime?: number;
-  item_address?: string;
-  limit?: number;
-  offset?: number;
-  sort?: "ASC" | "DESC";
-  start_lt?: number;
-  start_utime?: number;
-};
-
-type EstimateFeeParams = {
-  address?: string;
-  body?: string;
-  ignore_chksig?: boolean;
-  init_code?: string;
-  init_data?: string;
-};
-
-type GetAddressInformationParams = {
-  address: string;
-};
-
-type GetAddressStateParams = {
-  address: string;
-};
-
-type GetBlockHeaderParams = {
-  file_hash?: string;
-  root_hash?: string;
-  seqno: number;
-  shard: string;
-  workchain: number;
-};
-
-type GetBlockTransactionsParams = {
-  after_hash?: string;
-  after_lt?: number;
-  count?: number;
-  file_hash?: string;
-  root_hash?: string;
-  seqno?: number;
-  shard?: string;
-  workchain?: number;
-};
-
-type GetExtendedAddressInformationParams = {
-  address: string;
-};
-
-type GetMasterchainBlockSignaturesParams = {
-  seqno: number;
-};
-
-type GetTokenDataParams = {
-  address: string;
-};
-
-/**
- * @param address - TON address of the smart contract you want to execute the `get-method`
- * @param method - The function name of the `get-method`
- * @param stack - The input parameter of the `get-method` if it needs
- */
-type RunGetMethodParams = {
-  address: string;
-  method: string;
-  stack?: Array<["num", number] | ["cell", string] | ["slice", string]>;
-};
-
-/**
- * @description TON Center V3 style RunGetMethod parameters
- * @param address - TON address of the smart contract you want to execute the `get-method`
- * @param method - The function name of the `get-method`
- * @param stack - The input parameter of the `get-method` if it needs
- */
-type RunGetMethodV3Params = {
-  address: string;
-  method: string;
-  stack?: Array<{ type: "num" | "cell" | "slice"; value: string; }>;
-};
-
-type RunGetMethodResponse = {
-  exit_code: number;
-  stack: any[];
-  gas_used: number;
-  "@type": string;
-  "@extra": string;
-};
-
-type SendMessageParams = {
-  boc: string;
-};
-
-type GetBocStatusParams = {
-  boc: string;
-}
-
-type VerifyBocParams = {
-  boc: string;
-}
-
-type RadixConversionParams = {
-  base: string;
-  number: string;
-}
-
-type BinaryConversionParams = {
-  ascii: string;
-  base64?: string;
-  base64url?: string;
-  hexadecimal?: string;
-}
-
-type DetectAddressParams = {
-  address: string;
-}
-
-type GetTgBTCConfigResponse = {
-  jetton_master: string;
-  teleport: string;
-  coordinator: string;
-};
-
-/**
- * @param {string} address - TON address of the owner's Jetton Wallet (base64, base64Url, or hexadecimal)  
- */
-type GetTgBTCBalanceParams = {
-  address: string;
-};
-
-type GetTgBTCBalanceResponse = {
-  address: string;
-  balance: string;
-};
-
-type GetTgBTCMasterAddressResponse = {
-  address: string;
-  account_friendly: string;
-};
-
-/**
- * @param {number} limit - The number of tgBTC holders you want to see
- * @param {number} offset - The number of tgBTC holders skipped
- */
-type GetTgBTCHoldersParams = {
-  limit?: number;
-  offset?: number;
-};
-
-type GetTgBTCHoldersResponse = {
-  holders: {
-    address: string;
-    balance: string;
-    last_updated: number;
-    owner_type: string;
-  }[];
-  total: number;
-};
-
-type GetTgBTCBurnsAddress = { address: string; } | { jetton_wallet: string; };
-
-type GetTgBTCBurnsUtime = {
-  start_utime: number;
-  end_utime: number;
-} | {
-  start_utime?: undefined;
-  end_utime?: undefined;
-};
-
-type GetTgBTCBurnsLt = {
-  start_lt: number;
-  end_lt: number;
-} | {
-  start_lt?: undefined;
-  end_lt?: undefined;
-};
-
-/**
- * @param {string} address - TON address of the tgBTC owner (required when jetton_wallet is absent) (base64, base64Url, or hexadecimal)
- * @param {string} jetton_wallet - TON address of the tgBTC Jetton Wallet (required when address is absent) (base64, base64Url, or hexadecimal)
- * @param {number} start_utime - The beginning transaction Unix timestamp (required when end_utime is used)
- * @param {number} end_utime - The ending transaction Unix timestamp (required when start_utime is used)
- * @param {number} start_lt - The beginning transaction logical time (LT) (required when end_lt is used)
- * @param {number} end_lt - The ending transaction logical time (LT) (required when start_lt is used)
- * @param {"ASC" | "DESC"} sort - Enable sorting on responses by logical time (LT)
- * @param {number} limit - The number of tgBTC burn messages you want to see
- * @param {number} offset - The number of tgBTC burn messages skipped
- */
-type GetTgBTCBurnsParams = GetTgBTCBurnsAddress
-  & GetTgBTCBurnsUtime
-  & GetTgBTCBurnsLt
-  & Partial<{ sort: "ASC" | "DESC"; limit: number; offset: number; }>;
-
-type GetTgBTCBurnsResponse = {
-  query_id: string;
-  owner: string;
-  jetton_master: string;
-  jetton_wallet: string;
-  amount: string;
-  transaction_hash: string;
-  transaction_lt: string;
-  transaction_now: number;
-  response_destination?: string;
-  custom_payload?: string;
-}
-
-type GetTgBTCWalletAddressByOwnerParams = {
-  owner_address: string;
-}
-
-type GetTgBTCWalletAddressByOwnerResponse = {
-  address: string;
-  address_friendly: string;
-  owner: string;
-  owner_friendly: string;
-  jetton: string;
-  jetton_friendly: string;
-}
-
-type GetTgBTCTransferPayload = {
-  amount: number;
-  destination: string;
-  source: string;
-  comment?: string;
-}
-
-type GetTgBTCTransferResponse = {
-  address: string;
-  amount: number;
-  payload: string;
-}
-
-type TgBTCMetaDataResponse = {
-  address: string;
-  name: string;
-  symbol: string;
-  total_supply?: string;
-  admin_address?: string;
-  decimals?: string;
-  mintable?: boolean;
-  uri?: string;
-  image?: string;
-  image_data?: number[];
-  description?: string;
-}
-
-type TimeRange = {
-  start_utime: number;
-  end_utime: number;
-} | {
-  start_utime?: undefined;
-  end_utime?: undefined;
-};
-
-type LtRange = {
-  start_lt: number;
-  end_lt: number;
-} | {
-  start_lt?: undefined;
-  end_lt?: undefined;
-};
-
-type GetTgBTCTransfersParams = {
-  address: string;
-  jetton_wallet?: string;
-  direction?: 'in' | 'out' | 'both'
-  sort?: 'ASC' | 'DESC';
-  /** Use with limit to batch read (1-256) */
-  limit?: number;
-  /** Skip first N rows*/
-  offset?: number;
-
-} & TimeRange & LtRange
-
-type GetTgBTCTransfersResponse = {
-  query_id: string;
-  source: string;
-  destination: string;
-  amount: string;
-  source_wallet: string;
-  jetton_master: string;
-  transaction_hash: string;
-  transaction_lt: string;
-  transaction_now: number;
-  response_destination?: string;
-  custom_payload?: string;
-  forward_ton_amount: string;
-  forward_payload?: string;
-}
-
-type TONXRunAction =
-  | RunAction
-  | {
-    method: "getAccountBalance";
-    params: GetAccountBalanceParams;
-  }
-  | {
-    method: "getTransactions";
-    params: GetTransactionsParams;
-  }
-  | {
-    method: "getJettonBurns";
-    params: GetJettonBurnsParams;
-  }
-  | {
-    method: "getJettonMasters";
-    params: GetJettonMastersParams;
-  }
-  | {
-    method: "getJettonTransfers";
-    params: GetJettonTransfersParams;
-  }
-  | {
-    method: "getJettonWallets";
-    params: GetJettonWalletsParams;
-  }
-  | {
-    method: "getMessages";
-    params: GetMessagesParams;
-  }
-  | {
-    method: "getNftCollections";
-    params: GetNftCollectionsParams;
-  }
-  | {
-    method: "getNftItems";
-    params: GetNftItemsParams;
-  }
-  | {
-    method: "getNftTransfers";
-    params: GetNftTransfersParams;
-  }
-  | {
-    method: "estimateFee";
-    params: EstimateFeeParams;
-  }
-  | {
-    method: "getAddressInformation";
-    params: GetAddressInformationParams;
-  }
-  | {
-    method: "getAddressState";
-    params: GetAddressStateParams;
-  }
-  | {
-    method: "getBlockHeader";
-    params: GetBlockHeaderParams;
-  }
-  | {
-    method: "getBlockTransactions";
-    params: GetBlockTransactionsParams;
-  }
-  | {
-    method: "getConsensusBlock";
-  }
-  | {
-    method: "getExtendedAddressInformation";
-    params: GetExtendedAddressInformationParams;
-  }
-  | {
-    method: "getMasterchainBlockSignatures";
-    params: GetMasterchainBlockSignaturesParams;
-  }
-  | {
-    method: "getTokenData";
-    params: GetTokenDataParams;
-  }
-  | {
-    method: "runGetMethod";
-    params: RunGetMethodParams;
-  }
-  | {
-    method: "sendMessage";
-    params: SendMessageParams;
-  }
-  | {
-    method: "getBocStatus";
-    params: GetBocStatusParams;
-  } | {
-    method: "verifyBoc";
-    params: VerifyBocParams;
-  } | {
-    method: "radixConversion";
-    params: RadixConversionParams;
-  } | {
-    method: "binaryConversion";
-    params: BinaryConversionParams;
-  } | {
-    method: "detectAddress";
-    params: DetectAddressParams;
-  } | {
-    method: "getMasterchainInfo";
-  } | {
-    method: "getTgBTCConfig";
-  } | {
-    method: "getTgBTCBalance";
-    params: GetTgBTCBalanceParams;
-  } | {
-    method: "getTgBTCMasterAddress";
-  } | {
-    method: "getTgBTCHolders";
-    params: GetTgBTCHoldersParams;
-  } | {
-    method: "getTgBTCBurns";
-    params: GetTgBTCBurnsParams;
-  } | {
-    method: "getTgBTCTransferPayload",
-    params: GetTgBTCTransferPayload;
-  } | {
-    method: "getTgBTCTransfers",
-    params: GetTgBTCTransfersParams;
-  }
-
-export type TONXJsonRpcProviderOptions = JsonRpcApiProviderOptions & {
-  apiKey: string;
-  httpClientOptions?: CreateAxiosDefaults;
-};
 
 const version = "v2";
 export class TONXJsonRpcProvider extends JsonRpcProvider {
@@ -629,7 +158,7 @@ export class TONXJsonRpcProvider extends JsonRpcProvider {
       case "getTgBTCMasterAddress":
         return { method: "getTgBTCMasterAddress", params: {} };
       case "getTgBTCHolders":
-          return { method: "getTgBTCHolders", params: action.params };
+        return { method: "getTgBTCHolders", params: action.params };
       case "getTgBTCBurns":
         return { method: "getTgBTCBurns", params: action.params };
       case "getTgBTCWalletAddressByOwner":
